LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/try-tactic-percentage-expected.txt


PASS , left:anchor(0%), left:anchor(left)
PASS , left:anchor(100%), left:anchor(right)
PASS , left:anchor(50%), left:anchor(center)
PASS , top:anchor(0%), top:anchor(top)
PASS , top:anchor(100%), top:anchor(bottom)
PASS , top:anchor(50%), top:anchor(center)
FAIL flip-inline, left:anchor(10%), right:anchor(90%) assert_equals: offsetLeft expected 200 but got 120
FAIL flip-inline, left:anchor(calc(10% + 20%)), right:anchor(70%) assert_equals: offsetLeft expected 180 but got 140
PASS flip-block, left:anchor(0%), left:anchor(0%)
PASS flip-block, left:anchor(100%), left:anchor(100%)
PASS flip-block, top:anchor(0%), bottom:anchor(100%)
FAIL flip-block, top:anchor(100%), bottom:anchor(0%) assert_equals: offsetTop expected 110 but got 210
FAIL flip-inline, left:anchor(0%), right:anchor(100%) assert_equals: offsetLeft expected 210 but got 110
FAIL flip-inline, left:anchor(100%), right:anchor(0%) assert_equals: offsetLeft expected 110 but got 210
PASS flip-inline, top:anchor(0%), top:anchor(0%)
PASS flip-inline, top:anchor(100%), top:anchor(100%)
FAIL flip-block flip-inline, left:anchor(0%), right:anchor(100%) assert_equals: offsetLeft expected 210 but got 110
FAIL flip-block flip-inline, left:anchor(100%), right:anchor(0%) assert_equals: offsetLeft expected 110 but got 210
PASS flip-block flip-inline, top:anchor(0%), bottom:anchor(100%)
FAIL flip-block flip-inline, top:anchor(100%), bottom:anchor(0%) assert_equals: offsetTop expected 110 but got 210
FAIL flip-start, left:anchor(0%), top:anchor(0%) assert_equals: offsetLeft expected 0 but got 150
FAIL flip-start, left:anchor(100%), top:anchor(100%) assert_equals: offsetLeft expected 0 but got 250
FAIL flip-start, bottom:anchor(0%), right:anchor(0%) assert_equals: offsetLeft expected 110 but got 0
FAIL flip-start, bottom:anchor(100%), right:anchor(100%) assert_equals: offsetLeft expected 210 but got 0
FAIL flip-block flip-start, left:anchor(0%), top:anchor(0%) assert_equals: offsetLeft expected 0 but got 150
FAIL flip-block flip-start, left:anchor(100%), top:anchor(100%) assert_equals: offsetLeft expected 0 but got 250
FAIL flip-block flip-start, bottom:anchor(0%), left:anchor(100%) assert_equals: offsetLeft expected 250 but got 0
FAIL flip-block flip-start, bottom:anchor(100%), left:anchor(0%) assert_equals: offsetLeft expected 150 but got 0
FAIL flip-inline flip-start, left:anchor(0%), bottom:anchor(100%) assert_equals: offsetLeft expected 0 but got 110
FAIL flip-inline flip-start, left:anchor(100%), bottom:anchor(0%) assert_equals: offsetLeft expected 0 but got 210
FAIL flip-inline flip-start, bottom:anchor(0%), right:anchor(0%) assert_equals: offsetLeft expected 110 but got 0
FAIL flip-inline flip-start, bottom:anchor(100%), right:anchor(100%) assert_equals: offsetLeft expected 210 but got 0
FAIL flip-block flip-inline flip-start, left:anchor(0%), bottom:anchor(100%) assert_equals: offsetLeft expected 0 but got 110
FAIL flip-block flip-inline flip-start, left:anchor(100%), bottom:anchor(0%) assert_equals: offsetLeft expected 0 but got 210
FAIL flip-block flip-inline flip-start, bottom:anchor(0%), left:anchor(100%) assert_equals: offsetLeft expected 250 but got 0
FAIL flip-block flip-inline flip-start, bottom:anchor(100%), left:anchor(0%) assert_equals: offsetLeft expected 150 but got 0
